Add OTP to delete user

+++ controllers/user/deleteuser.js
@@ -2,10 +2,10 @@ import bcrypt from "bcryptjs";
 import supabase from "../database/database.js";
 
 export const deleteuser = async (req, res) => {
-    const { emailoruser, password } = req.body;
+    const { emailoruser, otp, password } = req.body;
     var { data, error } = await supabase
         .from("user_info")
-        .select('*')
+        .select('id, otp, password')
         .or(`email.eq.${emailoruser},username.eq.${emailoruser}`);
     if (error) {
         res.status(500).json({ error: true, message: error.message });
@@ -15,14 +15,14 @@ export const deleteuser = async (req, res) => {
         } else {
             const user = data[0];
             const validPassword = await bcrypt.compare(password, user.password);
-            if (validPassword) {
+            const validOTP = await bcrypt.compare(otp, user.otp);
+            if (validPassword && validOTP) {
                 var {data, error} = await supabase.from("user_info")
                     .delete()
                     .eq("id", user.id);
                 if (error) {
                     res.status(500).json({ error: true, message: error.message });
-                }
-                else {
+                } else {
                     res.status(200).json({ message: "user deleted" });
                 }
             } else {
